Move MatchFileOwnersCases to pkg/internal/utils
