+++ pkg/provider/gardener/internal/utils/utils.go
@@ -7,7 +7,6 @@ package utils
 import (
 	"context"
 	"fmt"
-	"slices"
 
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
@@ -122,35 +121,6 @@ func anyNodesForWorkerGroup(workerGroupName string, nodes []corev1.Node) bool {
 		}
 	}
 	return false
-}
-
-// MatchFileOwnersCases returns []rule.CheckResult for a given file and its owners for a select expected values.
-func MatchFileOwnersCases(
-	fileOwnerUser,
-	fileOwnerGroup,
-	fileName string,
-	expectedFileOwnerUsers,
-	expectedFileOwnerGroups []string,
-	target rule.Target,
-) []rule.CheckResult {
-	checkResults := []rule.CheckResult{}
-
-	if !slices.Contains(expectedFileOwnerUsers, fileOwnerUser) {
-		detailedTarget := target.With("details", fmt.Sprintf("fileName: %s, ownerUser: %s, expectedOwnerUsers: %v", fileName, fileOwnerUser, expectedFileOwnerUsers))
-		checkResults = append(checkResults, rule.FailedCheckResult("File has unexpected owner user", detailedTarget))
-	}
-
-	if !slices.Contains(expectedFileOwnerGroups, fileOwnerGroup) {
-		detailedTarget := target.With("details", fmt.Sprintf("fileName: %s, ownerGroup: %s, expectedOwnerGroups: %v", fileName, fileOwnerGroup, expectedFileOwnerGroups))
-		checkResults = append(checkResults, rule.FailedCheckResult("File has unexpected owner group", detailedTarget))
-	}
-
-	if len(checkResults) == 0 {
-		detailedTarget := target.With("details", fmt.Sprintf("fileName: %s, ownerUser: %s, ownerGroup: %s", fileName, fileOwnerUser, fileOwnerGroup))
-		checkResults = append(checkResults, rule.PassedCheckResult("File has expected owners", detailedTarget))
-	}
-
-	return checkResults
 }
 
 // MatchFilePermissionsAndOwnersCases returns []rule.CheckResult for a given file and its permissions and owners for a select expected values.
